NP-3206 Split files
