+++ src/pages/registration/resource_type_tab/ArtisticTypeForm.tsx
@@ -5,7 +5,7 @@ import { lightTheme } from '../../../themes/lightTheme';
 import { ArtisticType, ResourceFieldNames } from '../../../types/publicationFieldNames';
 import { ArtisticRegistration } from '../../../types/publication_types/artisticRegistration.types';
 import { SelectTypeField } from './components/SelectTypeField';
-import { ArtisticDesignForm } from './sub_type_forms/artistic_types/ArtisticDesignForm';
+import { ArtisticDesignForm } from './sub_type_forms/artistic_types/design/ArtisticDesignForm';
 
 interface ArtisticTypeFormProps {
   onChangeSubType: (type: string) => void;
